+++ src/main/java/io/github/yoshikawaa/app/githubmanager/web/entity/PullRequestReport.java
@@ -14,5 +14,5 @@ import lombok.Setter;
 public class PullRequestReport {
 	private Issue pull;
 	private Map<String, Integer> statusSummary;
-	private List<Integer> relatedIssueNumbers;
+	private List<Long> relatedIssueNumbers;
 }
